+++ config/activitylogs.php
@@ -1,0 +1,9 @@
+<?php
+
+return [
+    'except-routes' => [
+        'login',
+        'logout',
+    ],
+    'user' => 'id'
+];