+++ CHANGELOG.md
@@ -1,1 +1,16 @@
 # Changelog
+
+[11 Oct 2023]
+v0.0.3
+- added auto logging function
+- added config files
+
+
+[17 Feb 2023]
+v0.0.2
+- added support for laravel v10.x
+
+
+[06 Nov 2022]
+v0.0.1
+- first release